==> 11-evaluate-best-growth-models.R <==
#' Load data from multiple files to a list
#'
#' Intended for files generated by repeating an operation many times, saving
#' the results to an Rdata file each time. Each file is named for the name of
#' the single object it contains, and a number appended.
#'
#' @param a character vector containing file paths
#' @return a list containing the contents of each file, where each list
#'   element is named for the file it was loaded from (with the .Rdata
#'   extension stripped).
#'
load_to_list <- function(files) {
    e <- new.env()
    load_and_rename <- function(file, envir) {
        load(file, envir = envir)
        new_name <- gsub('-', '_', tools::file_path_sans_ext(basename(file)))
        old_name <- gsub('_[0-9]+$', '', new_name)
        assign(new_name, get(old_name), envir = envir)
        rm(list = old_name, envir = envir)
    }
    lapply(files, load_and_rename, envir = e)
    return(as.list(e))
}

# Load model comparison results ----
files <- file.path("network", list.files(path = "network", pattern = "model-comparisons-[0-9]{3}.Rdata"))
X <- load_to_list(files)

# Determine optimal log-likelihood for each model ----
L0 <- vapply(
    X,
    FUN = function(x, colvar) {x[, colvar]},
    FUN.VALUE = numeric(28),
    colvar = "L0"
)
L1 <- vapply(
    X,
    FUN = function(x, colvar) {x[, colvar]},
    FUN.VALUE = numeric(28),
    colvar = "L1"
)
L0_min <- apply(L0, 1, min)
L1_min <- apply(L1, 1, min)

# Recompute statistics for optimal model comparisons ----
all.equal(names(L0_min), names(L1_min))
chisq <- netgrowr:::likelihood_ratio(L1_min, L0_min)
df <- X[[1]][, "df"]
nobs <- X[[1]][, "nobs"]
p <- pchisq(chisq, df = df, lower.tail = FALSE)

optimal_model_comparisons <- data.frame(
    df = df,
    nobs = nobs,
    L0 = L0_min,
    L1 = L1_min,
    chisq = chisq,
    p = p,
    p_fdr = p.adjust(p, method = "fdr"),
    BIC = netgrowr:::bayesian_information_criterion(chisq, df, 547),
    row.names = names(L0_min)
)

# Save optimal model comparisons ----
save(optimal_model_comparisons, file = "network/optimal_model_comparisons.Rdata")
write.csv(optimal_model_comparisons, file = "network/optimal_model_comparisons.csv")
